0.2.0: Post to middle-tier for search
This goes with version 0.2.0 of ml-treehouse-node

There is still much work to be done to define the contract, but this
library now talks primarily to the middle-tier (rather than directly to
MarkLogic via a thin proxy). It only pushes through qtext and then gets
the passed-through MarkLogic search response, but this means that much
of the plumbing is in place and working.