Reorganize search state
There is an emerging searchQueryStateShape, which exists in two places:

1. preExecutedSearch - this possibly badly-named property tracks the
search state being compiled by the user, but which has not necessarily
been executed. In some cases, such a change should always provoke a
search, for example when paginating or probably when selecting a facet
value. However, this is not true for qtext change, eg. When a search
should always be launched, two actions should be generated: first, the
query change; second, the search.

2. executedSearch.query - This is more deeply nested because the
executedSearch also tracks results and bookkeeping such as the status of
the search execution.