Require consumer of runSearch to pass searchQuery
This allows all reducers listening for this action to see the
searchQuery without having to inspect other parts of the state. It also
allows runSearch to be invoked separately from the preExecutedSearch, so
long as the consumer passes some valid searchQuery. This should also
allow us to separate the preExecutedSearch reducer from the
executedSearch reducer, since neither needs access to the state of the
other. (The executedSearch reducer still has to be changed to make this
real, pulling the executed search from the action payload.)

There is another principle at work here: avoiding inspecting the state
from within an actionCreator, even a thunk, where `getState` is made
available. The problem with inspecting state is that this operates at a
global level, meaning that this duck needs to know somehow where it
lives in the global state tree. It would be better for modularity if it
did not need to know that.