Group search response info together
And provide more selectors into the state, to avoid the need for
consumers to ever need to know the shape of the state.

Make these selectors only depend on the local slice of state, so they
do not need to know where in the state tree they live. This does place
an additional burden on consuming code, which now has to wrap them like
so:

```javascript
import { searchSelectors } from 'ml-search-redux';
const wrappedSearchSelectors = Object.keys(searchSelectors).reduce(
  (newSelectors, name) => {
    newSelectors[name] = state => searchSelectors[name](state.search)
    return newSelectors;
  },
  {}
);
```

Also, begin to document the shape of the state using TypeScript-like
syntax.

And clean up the tests, and include some information on consuming this
library in the README.