Allow simple namespacing of reducers and actions
This allows this module to be consumed multiple times in a single app
without collision, by namespacing the actionTypes. Still experimenting
with the best way to handle this and looking around at state of the art.

Currently, you can import createReducer() and createActions() and pass
them an object like `{namespace: 'standard'}`. This, along with binding
the selector functions, allows a duck module to be mounted in multiple
locations.

One downside is that the consuming application has to do this twice,
once for the reducer and once for the actions