Maintain stale results while search pending
This means the old results are still available to the UI, in order to
avoid UI flickering when the old results are removed before new ones are
available.

Note that this could probably be improved, to allow the UI to query
whether results are stale, for example. Also, perhaps this should be
made available via search history instead. But, for now, simplicity
suggests leaving it as is.

Also, note that I removed some brittle reducer tests, adding an extra
integration test instead to document and ensure the continuance of this
behavior.